ECS GitAction/ecs-actions/requirements.txt
Flask==3.0.2
SQLAlchemy==2.0.29
PyMySQL==1.1.0
Werkzeug>=2.1 
python-dotenv==1.0.1
gunicorn==21.2.0
